requirements.txt
Flask>=2.0.1
SQLAlchemy>=2.0.0
nltk>=3.6.3
scikit-learn>=1.0.0
pandas>=1.3.3
numpy>=1.21.2
python-dotenv>=0.19.0
Flask-SQLAlchemy>=2.5.1
Flask-Login>=0.5.0
Flask-Mail>=0.9.1
Flask-WTF>=0.15.1
Werkzeug>=2.0.1
spacy>=3.1.3
psycopg2-binary>=2.9.1
python-jose>=3.3.0
email-validator>=1.1.3
gunicorn>=20.1.0
Flask-Limiter>=3.5.0
requests>=2.31.0

# RAG và LLM dependencies
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
openai>=1.0.0
google-generativeai>=0.3.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.2.0
coverage>=7.2.0

# API Documentation
flask-swagger-ui>=4.11.1
